+++ JKC_Fallguys/Assets/1_Scripts/Scene/03_Stage/FinalRoundAudioScheduler.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1ae13cac9c51a4e5f8e208e1106d8bcf
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
